services/control-plane/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-control-plane"
version = "0.1.0"
description = "Nexus SaaS control plane"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "python-jose[cryptography]>=3.4.0",
  "passlib[argon2]>=1.7.4",
  "httpx>=0.28.1",
  "redis>=6.0.0",
  "cryptography>=44.0.0",
  "email-validator>=2.2.0",
  "psycopg[binary]>=3.2.3",
  "alembic>=1.14.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
